rework from the ground up
