+++ assets/moths_fluttering_around_a_light_source/license.txt
@@ -1,0 +1,8 @@
+Model Information:
+* title:	🦋 moths fluttering around a light source 🦋
+* source:	https://sketchfab.com/3d-models/moths-fluttering-around-a-light-source-8d973cce8f654c20ae3f5208993d7e70
+* author:	Jazz Vincent (https://sketchfab.com/JazzVincent)
+
+Model License:
+* license type:	SKETCHFAB Standard (https://sketchfab.com/licenses)
+* requirements:	Under basic restrictions, use worldwide, on all types of media, commercially or not, and in all types of derivative works
